Convert RuleEditor to TS
